buscoescala_api/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
